+++ src/mikado.js
@@ -27,7 +27,7 @@ import {
     MIKADO_TPL_KEY,
     MIKADO_TPL_PATH,
     MIKADO_NODE_CACHE,
-    MIKADO_PROXY
+    MIKADO_PROXY, MIKADO_TPL_ROOT
 } from "./config.js";
 import { tick } from "./profiler.js";
 // <-- COMPILER BLOCK
@@ -1120,6 +1120,11 @@ Mikado.prototype.create = function(data, state, index, _update_pool){
 
             node[MIKADO_NODE_CACHE] = cache;
         }
+
+        if(SUPPORT_EVENTS){
+
+            node[MIKADO_TPL_ROOT] = 1;
+        }
     }
 
     if(keyed){
@@ -1166,45 +1171,6 @@ Mikado.prototype.add = function(data, state, index){
     }
 
     const node = this.create(data, state, index, 1);
-
-    //if(SUPPORT_STORAGE) {
-
-        //let stealth_mode;
-
-        // if(SUPPORT_REACTIVE && this.proxy){
-        //
-        //     if(this.stealth && this.loose && (node["_data"] === data)){
-        //
-        //         stealth_mode = 1;
-        //     }
-        //     else{
-        //
-        //         data["_proxy"] || (data = proxy_create(data, node["_path"] || this.create_path(node), this.proxy));
-        //     }
-        // }
-
-        // if(!stealth_mode){
-        //
-        //     if(this.store){
-        //
-        //         if(has_index && !this.extern){
-        //
-        //             splice(this.store, this.length - 1, length, data);
-        //             //this.store.splice(length, 0, data);
-        //         }
-        //         else{
-        //
-        //             if(SUPPORT_REACTIVE) this.skip = 1;
-        //             this.store[length] = data;
-        //             if(SUPPORT_REACTIVE) this.skip = 0;
-        //         }
-        //     }
-        //     else if(this.loose){
-        //
-        //         node["_data"] = data;
-        //     }
-        // }
-    //}
 
     if(has_index){